Move categoryName into Requested
This turned out to be a bit uglier than I expected because we have to
add it to the failure case too, to make sure we have a category name
in all cases. We also end up extracting the category name out of the
tracker, which feels like a violation of an abstraction.