Fully embrace the idea of imposing an order on requests
This is something of a substantial rewrite. We now require the user to
tell us how to order requests and we respect that order when updating
our state. This necessitates keeping the tracker around in success and
failure cases, but that's actually kind of a boon because we can use
the category name embedded in the tracker in the example app,
simplifying stuff a bit. The example is still kind of ugly, and I'm
starting to get uncomfortable exposing the constructors, but I don't
have a clearer idea of how else to let people pattern match
exhaustively, which is the whole point.